Source Code Uploaded!
Finally got around to uploading the source code. There are still things
to be done. Once a couple of things are fixed a released will be made.
In the mean time people that know how to compile the program can do so
but make sure to adhere to the license.